+++ mobile/android/android-components/components/concept/awesomebar/src/main/java/mozilla/components/concept/awesomebar/AwesomeBar.kt
@@ -122,7 +122,8 @@ interface AwesomeBar {
         enum class Flag {
             BOOKMARK,
             OPEN_TAB,
-            CLIPBOARD
+            CLIPBOARD,
+            SYNC_TAB
         }
 
         /**